No bugs in LocallyNameless, but in the proof.
Found that there was no bugs in ToLocallyNameless, but
that the proof bigTerm.sie had a couple of errors in
terms of free variables etc. I changed the lambda from
\ Var . Term1
to
\ Var . Term
This introduces two shift/reduce conflicts,
exemplified by
\x . x + y
(\x. x) + y (before)
\x. (x + y) (now)
and
\x. x y
(\x.x) y (before)
\x. (x y) (now)
As you see above, it makes more sense to shift than to
reduce in these cases, so the shift/reduce conflict
resulting in a shift is better than interpreting the
expressions in the way that the before-terms indicate.

The changes I did to ProofChecker and ToLocallyNameless
were just changing the output type of
ToLocallyNameless.
Since the maps for the bound variables will allways
be empty when the conversion is complete, it only makes
sense to send out the free variables.