Revised the checking and added check on bound vars in substitution. I revised the checking so that instead of dealing with the context around each law explicitly, I transformed each law M ~> N to C[M] ~> C[N] and substituted into that context instead. This way there is much less code and you can apply laws in multiple places at once. I didn't think this was possible before, since I thought that I had to rename the bound variables of a context, but since I apply the substitutions leaf and up, I will never get into that situation.
I removed dead code in ProofChecker that was to be removed because of this change.

I also checked that all bound variables in substitutions are distinct from each other, the free variables and the
variables that are substituted into a binding position. I think the check is neccesary. If nothing else, it ensures
the validity of the arguments going into runSubstM.